+++ familyos_ultrabert/__init__.py
@@ -41,7 +41,7 @@ Capabilities:
 License: Proprietary - All Rights Reserved
 """
 
-__version__ = "2.2.0"
+__version__ = "2.2.1"
 __author__ = "FamilyOS Team"
 
 from familyos_ultrabert.model import UltraBERT